+++ tests/integration/cases/float_special_values/Ada_combined.adb
@@ -1,21 +1,14 @@
+with A_Stub; use A_Stub;
 procedure Check is
-   procedure Check_Declaration is
-      my_data : A_Val := AList'(
-          AFloat (1.0 / 0.0),
-          AFloat (-1.0 / 0.0),
-          AFloat (0.0 / 0.0)
-      );
-   begin
-      null;
-   end Check_Declaration;
-   procedure Check_Assignment is
-   begin
-      my_data := AList'(
-          AFloat (1.0 / 0.0),
-          AFloat (-1.0 / 0.0),
-          AFloat (0.0 / 0.0)
-      );
-   end Check_Assignment;
+   my_data : A_Val := AList'[
+       AFloat (1.0 / 0.0),
+       AFloat (-1.0 / 0.0),
+       AFloat (0.0 / 0.0)
+   ];
 begin
-   null;
+   my_data := AList'[
+       AFloat (1.0 / 0.0),
+       AFloat (-1.0 / 0.0),
+       AFloat (0.0 / 0.0)
+   ];
 end Check;