+++ tests/integration/cases/float_special_values/Ada_float_format_fixed_v.adb
@@ -1,9 +1,10 @@
+with A_Stub; use A_Stub;
 procedure Check is
-   my_data : A_Val := AList'(
+   my_data : A_Val := AList'[
        AFloat (1.0 / 0.0),
        AFloat (-1.0 / 0.0),
        AFloat (0.0 / 0.0)
-   );
+   ];
 begin
    null;
 end Check;